requirements.txt
yt-dlp>=2024.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
Pillow>=10.0.0
moviepy>=1.0.3
pysrt>=1.1.2
customtkinter>=5.2.0
CTkMessagebox>=2.5
